Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,23 +1,23 @@
 #include<stdio.h>
 int main()
 {
-    float a,b;
+    int a,b;
     char op;
     printf("");
-    scanf("%f %f %c",&a,&b,&op);
+    scanf("%d %d %c",&a,&b,&op);
     switch(op)
     {
         case '+':
-            printf("%f\n",a+b);
+            printf("%d\n",a+b);
             break;
         case '-':
-            printf("%f\n",a-b);
+            printf("%d\n",a-b);
             break;
         case '*':
-            printf("%f\n",a*b);
+            printf("%d\n",a*b);
             break;
         case '/':
-                printf("%f\n",a/b);
+                printf("%d\n",a/b);
                 break;
         default:
                 printf("Error\n");
